Reimpl with while loops and dictionary, success
This reimplements my logic using while loops, so I can control the
index, search ahead for a full number, update the index to beyond the
number, then continue on in the main loop.

Storing of parts is done using coordinates of the parts as an index into
a dictionary, to prevent duplicating particular parts in the schematic.