Keep track of the location of all ints and symbols
Create a class called SchematicItem that holds the state of an item in
the schematic -- that is, anything that is not a `.`.

If a digit is encountered, all consecutive digits are concatenated to
form an integer, and a single SchematicItem is created for that single
integer. A SchematicItem is not created for individual digits, unless
there is only a single consecutive digit found between other symbols.

This should facilitate answering some other followup questions.